Adding POST visit functionality

+++ src/main/java/com/example/docvisitorregister/domain/dao/Doctor.java
@@ -33,5 +33,5 @@ public class Doctor {
     private ZoneOffset timeZone;
 
     @Column(nullable = false)
-    private int totalPatients;
+    private Long totalPatients;
 }
