+++ src/main/java/com/example/docvisitorregister/domain/dto/DoctorDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class DoctorDTO {
     String firstName;
     String lastName;
-    Integer totalPatients;
+    Long totalPatients;
 }
